Decompose loom-daemon init.rs into focused submodules
Split the 2,300+ line init.rs monolith into 6 focused submodules:

- mod.rs (566 lines): Orchestration, public types, InitReport/ValidationReport,
  and integration tests for the initialize_workspace() entry point
- scaffolding.rs (760 lines): Repository scaffolding setup including CLAUDE.md
  preservation logic, .claude/, .codex/, .github/ copying, template substitution
- file_ops.rs (601 lines): File copy/merge/clean/verify operations with
  TemplateContext for avoiding false-positive verification failures
- git.rs (362 lines): Git detection (is_loom_source_repo), validation, repo info
  extraction, defaults path resolution
- templates.rs (121 lines): LoomMetadata and substitute_template_variables()
- post_init.rs (97 lines): Manifest generation and gitignore updates

Key improvements:
- Each domain is self-contained with high internal cohesion
- Tests are co-located with their associated logic (26 total tests preserved)
- Single public entry point preserved: init::initialize_workspace()
- Added TemplateContext support for template-aware file verification (from main)

Closes #1888
